pyproject.toml
[tool.poetry]
name = "smart-route-backend-api"
version = "0.1.0"
description = ""
authors = ["Dhrumil Mistry <56185972+dmdhrumilmistry@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "smart_route_backend_api"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
folium = "^0.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
